FIXED: converted Artemis 'int' entity-id to an 'NSUInteger' (required for Apple 64bit code), and typedef'd EntityID = NSUInteger (so that you don't need to remember this in your code - simply use EntityID where you mean an-integer-that-is-an-ID-for-an-entity)

+++ Artemis-Lib/Artemis-Lib/Classes/ArtemisBag.h
@@ -14,7 +14,7 @@
 
 -(id)initWithCapacity:(NSInteger) c;
 
--(NSObject*) remove:(NSInteger) index;
+-(NSObject*) remove:(NSUInteger) index;
 -(NSObject*) removeLast;
 /** Have to rename method, Objc doesn't support overloading, it's too basic/weak a language */
 -(BOOL) removeFirst:(id) item;
@@ -23,20 +23,20 @@
 
 -(BOOL) removeAll:(ArtemisBag*) otherBag;
 
--(NSObject*) get:(NSInteger) index;
+-(NSObject*) get:(NSUInteger) index;
 
-@property(nonatomic,readonly) NSInteger size, capacity;
+@property(nonatomic,readonly) NSUInteger size, capacity;
 
--(BOOL) isIndexWithinBounds:(int) bounds;
+-(BOOL) isIndexWithinBounds:(NSUInteger) bounds;
 
 @property(nonatomic,readonly) BOOL isEmpty;
 
 -(void) add:(NSObject*) item;
 
 /** Have to rename for objc */
--(void) setItem:(NSObject*) item atIndex:(int) index;
+-(void) setItem:(NSObject*) item atIndex:(NSUInteger) index;
 
--(void) ensureCapacity:(NSInteger) index;
+-(void) ensureCapacity:(NSUInteger) index;
 
 -(void) clear;
